Modified Trainings

+++ WebApplication/Controllers/KategoriaTreninguController.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +12,7 @@ using WebApplication.Models;
 
 namespace WebApplication.Controllers
 {
+    [Authorize]
     public class KategoriaTreninguController : Controller
     {
         private readonly MyContext _context;
@@ -33,8 +36,23 @@ namespace WebApplication.Controllers
                 return NotFound();
             }
 
-            var kategoriaTreningu = await _context.kategoriaTreningu
+            var kategoriaTreningu = await _context.kategoriaTreningu.Include(k => k.treningi)
                 .FirstOrDefaultAsync(m => m.id_kategorii == id);
+
+            ViewBag.trainings = kategoriaTreningu.treningi;
+            ViewBag.userId = int.Parse(User.Identity.GetUserId());
+
+            Rola usersRole = _context.role.Include(k => k.uzytkownicy)
+                                          .FirstOrDefault(m => m.nazwa == "trener");
+
+            List<int> trainersIds = new List<int>();
+            foreach(var user in usersRole.uzytkownicy)
+            {
+                trainersIds.Add(user.id_uzytkownika);
+            }
+
+            ViewBag.trainersIds = trainersIds;
+
             if (kategoriaTreningu == null)
             {
                 return NotFound();
@@ -46,6 +64,9 @@ namespace WebApplication.Controllers
         // GET: KategoriaTreningu/Create
         public IActionResult Create()
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             return View();
         }
 
@@ -73,6 +94,9 @@ namespace WebApplication.Controllers
                 return NotFound();
             }
 
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             var kategoriaTreningu = await _context.kategoriaTreningu.FindAsync(id);
             if (kategoriaTreningu == null)
             {
@@ -92,6 +116,9 @@ namespace WebApplication.Controllers
             {
                 return NotFound();
             }
+
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
 
             if (ModelState.IsValid)
             {
@@ -119,6 +146,9 @@ namespace WebApplication.Controllers
         // GET: KategoriaTreningu/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             if (id == null)
             {
                 return NotFound();
@@ -139,6 +169,22 @@ namespace WebApplication.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
+            KategoriaTreningu defaultCategory = _context.kategoriaTreningu
+                                                         .Where(k => k.nazwa == "inne")
+                                                         .FirstOrDefault();
+            if (defaultCategory.id_kategorii == id)
+                return RedirectToAction("Index");
+
+            List<Trening> trainings = _context.treningi.Where(k => k.id_kategorii == id).ToList();
+
+            foreach (Trening training in trainings)
+                training.id_kategorii = defaultCategory.id_kategorii;
+
+            _context.SaveChanges();
+
             var kategoriaTreningu = await _context.kategoriaTreningu.FindAsync(id);
             _context.kategoriaTreningu.Remove(kategoriaTreningu);
             await _context.SaveChangesAsync();
@@ -148,6 +194,17 @@ namespace WebApplication.Controllers
         private bool KategoriaTreninguExists(int id)
         {
             return _context.kategoriaTreningu.Any(e => e.id_kategorii == id);
+        }
+
+        private bool isTrainer()
+        {
+            int userId = int.Parse(User.Identity.GetUserId());
+            List<RolaUzytkownika> usersRoles = _context.RolaUzytkownika.Where(k => k.id_uzytkownika == userId).Include(c => c.rola).ToList();
+
+            foreach (var usersRole in usersRoles)
+                if (usersRole.rola.nazwa == "trener" || usersRole.rola.nazwa == "admin")
+                    return true;
+            return false;
         }
     }
 }
